requirements.txt
# # =========================
# # Core App
# # =========================
# streamlit==1.33.0
# python-dotenv==1.0.1

# # =========================
# # Data Handling (Python 3.12 compatible)
# # =========================
# numpy==1.26.4
# pandas==2.2.3

# # =========================
# # Machine Learning / Analytics
# # =========================
# scikit-learn==1.4.2
# scipy==1.12.0

# # =========================
# # LLM / API Clients
# # =========================
# groq==0.37.0
# openai==1.30.0
# httpx==0.27.0

# # =========================
# # Document Processing / Reports
# # =========================
# python-docx==1.1.0
# reportlab==4.0.9
# openpyxl==3.1.2

# # =========================
# # Visualization
# # =========================
# matplotlib==3.8.4


streamlit==1.33.0
python-dotenv==1.0.1

numpy==1.26.4
pandas==2.2.3

scikit-learn==1.4.2
scipy==1.12.0

groq==0.37.0
openai==1.30.0
httpx==0.27.0

# Document Processing
pdfplumber==0.11.4
pillow==10.4.0
python-docx==1.1.0
reportlab==4.0.9
openpyxl==3.1.2

matplotlib==3.8.4
